Moved listener-related code to a `host::descriptor::listener` module
