+++ src/main/host/descriptor/epoll/mod.rs
@@ -8,10 +8,8 @@ use linux_api::errno::Errno;
 use linux_api::ioctls::IoctlRequest;
 use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 
-use crate::host::descriptor::{
-    File, FileMode, FileSignals, FileState, FileStatus, StateEventSource, StateListenHandle,
-    StateListenerFilter, SyscallResult,
-};
+use crate::host::descriptor::listener::{StateEventSource, StateListenHandle, StateListenerFilter};
+use crate::host::descriptor::{File, FileMode, FileSignals, FileState, FileStatus, SyscallResult};
 use crate::host::memory_manager::MemoryManager;
 use crate::host::syscall::io::IoVec;
 use crate::host::syscall::types::SyscallError;
