+++ src/main/host/descriptor/socket/inet/legacy_tcp.rs
@@ -11,11 +11,11 @@ use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 
 use crate::core::worker::Worker;
 use crate::cshadow as c;
+use crate::host::descriptor::listener::{StateListenHandle, StateListenerFilter};
 use crate::host::descriptor::socket::inet::{self, InetSocket};
 use crate::host::descriptor::socket::{RecvmsgArgs, RecvmsgReturn, SendmsgArgs, Socket};
 use crate::host::descriptor::{
-    CompatFile, File, FileMode, FileSignals, FileState, FileStatus, OpenFile, StateListenHandle,
-    StateListenerFilter, SyscallResult,
+    CompatFile, File, FileMode, FileSignals, FileState, FileStatus, OpenFile, SyscallResult,
 };
 use crate::host::host::Host;
 use crate::host::memory_manager::MemoryManager;
