+++ src/main/host/descriptor/socket/inet/tcp.rs
@@ -13,12 +13,12 @@ use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 use crate::core::work::task::TaskRef;
 use crate::core::worker::Worker;
 use crate::cshadow as c;
+use crate::host::descriptor::listener::{StateEventSource, StateListenHandle, StateListenerFilter};
 use crate::host::descriptor::socket::inet;
 use crate::host::descriptor::socket::{InetSocket, RecvmsgArgs, RecvmsgReturn, SendmsgArgs};
 use crate::host::descriptor::{File, Socket};
 use crate::host::descriptor::{
-    FileMode, FileSignals, FileState, FileStatus, OpenFile, StateEventSource, StateListenHandle,
-    StateListenerFilter, SyscallResult,
+    FileMode, FileSignals, FileState, FileStatus, OpenFile, SyscallResult,
 };
 use crate::host::memory_manager::MemoryManager;
 use crate::host::network::interface::FifoPacketPriority;
